+++ Sources/FeaturePulse/Utilities/Color+Hex.swift
@@ -7,12 +7,16 @@ extension Color {
         hexSanitized = hexSanitized.hasPrefix("#") ? String(hexSanitized.dropFirst()) : hexSanitized
 
         guard hexSanitized.count == 6,
-              let rgb = UInt64(hexSanitized, radix: 16) else {
+              let rgb = UInt64(hexSanitized, radix: 16)
+        else {
             return nil
         }
 
+        // swiftlint:disable:next identifier_name
         let r = Double((rgb >> 16) & 0xFF) / 255.0
+        // swiftlint:disable:next identifier_name
         let g = Double((rgb >> 8) & 0xFF) / 255.0
+        // swiftlint:disable:next identifier_name
         let b = Double(rgb & 0xFF) / 255.0
 
         self.init(red: r, green: g, blue: b)